+++ src/conf_io/conf_obj.h
@@ -79,6 +79,15 @@ conf_doc_t* conf_obj_get_doc(object_t* conf);
  */
 ret_t conf_obj_save(object_t* conf);
 
+/*保存命令*/
+#define CONF_CMD_SAVE "save"
+
+/*和前一个兄弟节点交换位置*/
+#define CONF_CMD_MOVE_UP "move_up"
+
+/*和后一个兄弟节点交换位置*/
+#define CONF_CMD_MOVE_DOWN "move_down"
+
 END_C_DECLS
 
 #endif /*TK_CONF_OBJ_H*/