+++ tests/conf_ini_test.cc
@@ -271,3 +271,73 @@ TEST(Ini, file) {
 
   OBJECT_UNREF(conf);
 }
+
+TEST(Ini, index) {
+  value_t v;
+  conf_node_t* node = NULL;
+  conf_doc_t* doc = conf_doc_load_ini("[hello]\n[ world ]\n[awtk]\nname=aaa\n");
+
+  ASSERT_EQ(conf_doc_get(doc, "hello.#index", &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 0);
+  
+  ASSERT_EQ(conf_doc_get(doc, "world.#index", &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 1);
+
+  ASSERT_EQ(conf_doc_get(doc, "awtk.#index", &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 2);
+
+  conf_doc_destroy(doc);
+}
+
+TEST(Ini, last_first) {
+  value_t v;
+  conf_node_t* node = NULL;
+  conf_doc_t* doc = conf_doc_load_ini("[hello]\n[ world ]\n[awtk]\nname=aaa\n");
+
+  ASSERT_EQ(conf_doc_is_last(doc, "hello"), FALSE);
+  ASSERT_EQ(conf_doc_is_last(doc, "world"), FALSE);
+  ASSERT_EQ(conf_doc_is_last(doc, "awtk"), TRUE);
+  
+  ASSERT_EQ(conf_doc_is_first(doc, "hello"), TRUE);
+  ASSERT_EQ(conf_doc_is_first(doc, "world"), FALSE);
+  ASSERT_EQ(conf_doc_is_first(doc, "awtk"), FALSE);
+
+  conf_doc_destroy(doc);
+}
+
+TEST(Ini, move_up) {
+  value_t v;
+  conf_node_t* node = NULL;
+  conf_doc_t* doc = conf_doc_load_ini("[hello]\n[ world ]\n[awtk]\nname=aaa\n");
+
+  ASSERT_EQ(conf_doc_move_up(doc, "hello"), RET_FAIL);
+  ASSERT_EQ(conf_doc_is_first(doc, "hello"), TRUE);
+  ASSERT_EQ(conf_doc_get(doc, "hello.#index", &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 0);
+
+  ASSERT_EQ(conf_doc_move_down(doc, "hello"), RET_OK);
+  ASSERT_EQ(conf_doc_is_first(doc, "hello"), FALSE);
+  ASSERT_EQ(conf_doc_is_first(doc, "world"), TRUE);
+  ASSERT_EQ(conf_doc_get(doc, "hello.#index", &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 1);
+  
+  ASSERT_EQ(conf_doc_move_down(doc, "hello"), RET_OK);
+  ASSERT_EQ(conf_doc_is_last(doc, "hello"), TRUE);
+  ASSERT_EQ(conf_doc_get(doc, "hello.#index", &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 2);
+  ASSERT_EQ(conf_doc_move_down(doc, "hello"), RET_FAIL);
+  
+  ASSERT_EQ(conf_doc_move_up(doc, "hello"), RET_OK);
+  ASSERT_EQ(conf_doc_is_last(doc, "hello"), FALSE);
+  ASSERT_EQ(conf_doc_get(doc, "hello.#index", &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 1);
+  
+  ASSERT_EQ(conf_doc_move_up(doc, "hello"), RET_OK);
+  ASSERT_EQ(conf_doc_is_last(doc, "hello"), FALSE);
+  ASSERT_EQ(conf_doc_get(doc, "hello.#index", &v), RET_OK);
+  ASSERT_EQ(value_int(&v), 0);
+  
+  ASSERT_EQ(conf_doc_move_up(doc, "hello"), RET_FAIL);
+
+  conf_doc_destroy(doc);
+}